refactor: Reorganize webview and provider directories
- Rename src/views/ to src/providers/ (all *ViewProvider.ts files)
- Flatten src/webview/{dashboard,beads-panel,details,graph,kanban}/ to src/webview/views/
- Rename components to match View suffix convention:
  - Dashboard → DashboardView
  - BeadsPanel → IssuesView
  - BeadDetails → DetailsView
  - DependencyGraphView → GraphView
  - KanbanBoard → KanbanView
- Update all imports
