+++ src/providers/BeadDetailsViewProvider.ts
