+++ src/providers/BeadsPanelViewProvider.ts
